Remove deprecated DisplayManual class Move OmniRig to deprecated group, scheduled for deprecation due to being no longer supported by Windows 11 and Qt6.9.3 Schedule OmniRig to removed from the code in July 2026 Noted in the doxy file description
