+++ CMakeLists.txt
@@ -214,7 +214,6 @@ target_sources(
   JS8_Audio/NotificationAudio.cpp
   JS8_Audio/SoundInput.cpp
   JS8_Audio/SoundOutput.cpp
-  JS8_Deprecated/DisplayManual.cpp
   JS8_JSC/JSC.cpp
   JS8_JSC/JSC_checker.cpp
   JS8_JSC/JSC_list.cpp
@@ -452,7 +451,7 @@ if (WIN32)
       )
 
       target_sources(${TARGET} PRIVATE
-        JS8_Transceiver/OmniRigTransceiver.cpp
+        JS8_Deprecated/OmniRigTransceiver.cpp
         ${GENAXSRCS}
       )
